Import ohai version 5.2.2
