--- lib/provisioner/worker/plugins/automators/chef_solo_automator/resources/cookbooks/ohai/.foodcritic
@@ -1,2 +1,0 @@
-~FC016
-~FC009
