+++ lib/provisioner/worker/plugins/automators/chef_solo_automator/resources/cookbooks/ohai/resources/hint.rb
@@ -8,14 +8,14 @@ action :create do
     recursive true
   end
 
-  file ohai_hint_path do
+  file ohai_hint_file_path(new_resource.hint_name) do
     action :create
-    content build_content
+    content format_content(new_resource.content)
   end
 end
 
 action :delete do
-  file ohai_hint_path do
+  file ohai_hint_file_path(new_resource.hint_name) do # ~FC009
     action :delete
     notifies :reload, ohai[reload ohai post hint removal]
   end
@@ -26,24 +26,7 @@ action :delete do
 end
 
 action_class do
-  def ohai_hint_path
-    path = ::File.join(::Ohai::Config.ohai.hints_path.first, new_resource.hint_name)
-    path << '.json' unless path.end_with?('.json')
-    path
-  end
-
-  def build_content
-    # passing nil to file produces deprecation warnings so pass an empty string
-    return nil if new_resource.content.nil? || new_resource.content.empty?
-    JSON.pretty_generate(new_resource.content)
-  end
-
-  def file_content(path)
-    return JSON.parse(::File.read(path))
-  rescue JSON::ParserError
-    Chef::Log.debug("Could not parse JSON in ohai hint at #{ohai_hint_path}. It's probably an empty hint file")
-    return nil
-  end
+  include OhaiCookbook::HintHelpers
 end
 
 # this resource forces itself to run at compile_time
